requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi==0.70.0
uvicorn==0.15.0
jinja2
requests==2.26.0
sendgrid
SQLAlchemy
pymysql
pyodbc